+++ chaosgcp/__init__.py
@@ -37,7 +37,7 @@ __all__ = [
     "context_from_parent_path",
     "parse_interval",
 ]
-__version__ = "0.7.0"
+__version__ = "0.8.0"
 
 
 def get_service(