+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## [Unreleased][]
 
-[Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-google-cloud-platform/compare/0.7.0...HEAD
+[Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-google-cloud-platform/compare/0.8.0...HEAD
+
+## [0.8.0][] - 2023-02-07
+
+[0.8.0]: https://github.com/chaostoolkit-incubator/chaostoolkit-google-cloud-platform/compare/0.7.0...0.8.0
 
 ### Added